# apidocs/swagger.json
{
  "consumes": [
    "application/json",
    "text/plain"
  ],
  "produces": [
    "application/json",
    "text/plain"
  ],
  "schemes": [
    "https",
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This documentation describes the Supergiant Control API.",
    "title": "Supergiant Control API.",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "2.1.2"
  },
  "basePath": "/",
  "paths": {
...
    "/v1/api/tasks/{taskID}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get a task.",
        "operationId": "getTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TaskID",
            "name": "taskID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/taskResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/v1/api/tasks/{taskID}/logs": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Set up a websocket connection for logs streaming.",
        "operationId": "streamLogs",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TaskID",
            "name": "taskID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/v1/api/tasks/{taskID}/restart": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Restart a task.",
        "operationId": "restartTask",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TaskID",
            "name": "taskID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "$ref": "#/responses/emptyResponse"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorCode": {
      "type": "integer",
      "format": "int64",
      "x-go-package": "github.com/supergiant/control/pkg/sgerrors"
    },
...
    "Message": {
      "description": "Message is used in response body to display separate messages to user and developers",
      "type": "object",
      "properties": {
        "devMessage": {
          "description": "DevMessage should be used to display message to developer on API/console",
          "type": "string",
          "x-go-name": "DevMessage"
        },
        "errorCode": {
          "$ref": "#/definitions/ErrorCode"
        },
        "moreInfo": {
          "description": "MoreInfo should be a link to supergiant documentation to display common problems",
          "type": "string",
          "x-go-name": "MoreInfo"
        },
        "userMessage": {
          "description": "UserMessage should be used to display message to user on UI/CLI",
          "type": "string",
          "x-go-name": "UserMessage"
        }
      },
      "x-go-package": "github.com/supergiant/control/pkg/message"
    },
...
  },
  "responses": {
...
    "emptyResponse": {
      "description": "emptyResponse represents an empty http response."
    },
    "errorResponse": {
      "description": "errorResponse represents an error response.",
      "schema": {
        "$ref": "#/definitions/Message"
      }
    },
...
    "taskResponse": {
      "description": "taskResponse contains representations of a workflow task."
    },
...
  },
  "securityDefinitions": {
    "AuthorizationHeaderToken": {
      "description": "This scheme uses the header \"Authorization: Bearer {JWT token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}